Service with all status

+++ src/main/java/org/jibe77/hermanas/controller/music/MusicController.java
@@ -1,5 +1,8 @@
 package org.jibe77.hermanas.controller.music;
 
+import org.jibe77.hermanas.controller.ProcessLauncher;
+import org.jibe77.hermanas.controller.energy.SoundCardController;
+import org.jibe77.hermanas.scheduler.sun.ConsumptionModeManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
@@ -7,6 +10,8 @@ import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
+import javax.annotation.PostConstruct;
+import javax.annotation.PreDestroy;
 import java.io.*;
 import java.security.SecureRandom;
 import java.util.*;
@@ -46,78 +51,103 @@ public class MusicController {
     @Value("${music.volume.regular}")
     private String volumeLevelRegular;
 
+    @Value("${music.security.timer.delay.eco}")
+    private long musicSecurityTimerDelayEco;
+
+    @Value("${music.security.timer.delay.regular}")
+    private long musicSecurityTimerDelayRegular;
+
+    @Value("${music.security.timer.delay.sunny}")
+    private long musicSecurityTimerDelaySunny;
+
+    @Value("${music.enabled}")
+    private boolean musicEnabled;
+
     ProcessLauncher processLauncher;
 
     private Process currentMusicProcess;
 
+    private ConsumptionModeManager consumptionModeManager;
+
+    private Timer musicSecurityStopTimer;
+
+    private SoundCardController soundCardController;
+
     Logger logger = LoggerFactory.getLogger(MusicController.class);
 
-    public MusicController(ProcessLauncher processLauncher) {
+    public MusicController(ProcessLauncher processLauncher, ConsumptionModeManager consumptionModeManager,
+                           SoundCardController soundCardController) {
         this.processLauncher = processLauncher;
+        this.consumptionModeManager = consumptionModeManager;
+        this.soundCardController = soundCardController;
     }
 
     public boolean playMusicRandomly() {
-        stop();
-        try {
-            setMusicLevel(volumeLevelRegular);
-            List<String> listOfFile = getListOfFiles(pathToFolder);
-            playMusic(listOfFile);
-        } catch (IOException e) {
-            logger.error("Can't play music.", e);
+        if (musicEnabled) {
+            try {
+                stop();
+                setMusicLevel(volumeLevelRegular);
+                List<String> listOfFile = getListOfFiles(pathToFolder);
+                playMusic(listOfFile);
+            } catch (IOException e) {
+                logger.error("Can't play music.", e);
+                return false;
+            }
+            return true;
+        } else {
             return false;
         }
-        return true;
     }
 
-    private void playMusic(File musicFile) throws IOException {
+    private void playMusic(File musicFile, long duration) throws IOException {
         List<String> listOfFile = new ArrayList<>(1);
         listOfFile.add(musicFile.getAbsolutePath());
-        playMusic(listOfFile);
+        playMusic(listOfFile, duration);
     }
 
     private void playMusic(List<String> listOfFile) throws IOException {
+        playMusic(listOfFile, -1L);
+    }
+
+    private void playMusic(List<String> listOfFile, long duration) throws IOException {
         logger.info("Play music with command {} {} {}  {}.",
                 musicPlayerStartCmd, musicPlayerNoDispParam
                 , musicPlayerShuffle, listOfFile);
-        List<String> commandWithParams = new ArrayList<>(listOfFile.size() + 2);
+        soundCardController.turnOn();
+        List<String> commandWithParams = new ArrayList<>(listOfFile.size() + 3);
         commandWithParams.add(musicPlayerStartCmd);
         commandWithParams.add(musicPlayerNoDispParam);
         commandWithParams.add(musicPlayerShuffle);
         commandWithParams.addAll(listOfFile);
         currentMusicProcess = processLauncher.launch(commandWithParams);
-        printErrorStreamInThread(currentMusicProcess);
-    }
-
-    private void printErrorStreamInThread(Process currentMusicProcess) {
-        InputStream errorStream = currentMusicProcess.getErrorStream();
-        if (errorStream != null) {
-            BufferedReader bufferedReader = new BufferedReader(
-                    new InputStreamReader(errorStream));
-            new Thread(() -> {
-                String line = null;
-                logger.info("error stream is opened (debug only)...");
-                do {
-                    try {
-                        line = bufferedReader.readLine();
-                        if (line != null) {
-                            logger.debug(line);
-                        }
-                    } catch (IOException e) {
-                        logger.error("can't read process errors.", e);
-                    }
-                } while (line != null);
-                logger.info("process error stream is finished (debug only).");
-            }).start();
-        } else {
-            logger.info("error stream is null.");
-        }
+        processLauncher.printErrorStreamInThread(currentMusicProcess);
+        startSecurityTimer(duration);
     }
 
     private List<String> getListOfFiles(String pathToFolder) {
         File folder = new File(pathToFolder);
-        List<File> filesList = Arrays.asList(folder.listFiles());
+        File[] files = folder.listFiles();
+        List<File> filesList = files != null ? Arrays.asList(files) : Collections.emptyList();
         return filesList.stream()
                 .map(File::getAbsolutePath).collect(Collectors.toList());
+    }
+
+    private void startSecurityTimer(long durationParam) {
+        if (musicSecurityStopTimer != null) {
+            musicSecurityStopTimer.cancel();
+        }
+        musicSecurityStopTimer = new Timer("Music security stop");
+        final long duration = durationParam >= 0 ?
+                durationParam :
+                consumptionModeManager.getDuration(
+                    musicSecurityTimerDelayEco, musicSecurityTimerDelayRegular, musicSecurityTimerDelaySunny);
+        musicSecurityStopTimer.schedule(new TimerTask() {
+                                            public void run() {
+                                                logger.info("stopping music after {} ms.", duration);
+                                                stop();
+                                            }
+                                        },
+                duration);
     }
 
     private File pickSong(File[] array) {
@@ -126,34 +156,54 @@ public class MusicController {
     }
 
     public void stop() {
-        if (currentMusicProcess != null) {
+        if (musicEnabled && currentMusicProcess != null) {
             logger.info("Stop music destroying process.");
             currentMusicProcess.destroyForcibly();
             currentMusicProcess = null;
+            if (musicSecurityStopTimer != null) {
+                musicSecurityStopTimer.cancel();
+                musicSecurityStopTimer = null;
+            }
+            soundCardController.turnOff();
         }
     }
 
     public boolean cocorico() {
-        logger.info("Play cocorico !");
-        stop();
-        try {
-            setMusicLevel(volumeLevelMax);
-            File mixFolder = new File(pathToRooster);
-            File[] filesAvailable = mixFolder.listFiles();
-            File pickedFile = pickSong(filesAvailable);
-            playMusic(pickedFile);
-            return true;
-        } catch (IOException e) {
-            logger.error("Can't play cocorico.", e);
+        if (musicEnabled) {
+            logger.info("Play cocorico !");
+            stop();
+            try {
+                setMusicLevel(volumeLevelMax);
+                File mixFolder = new File(pathToRooster);
+                File[] filesAvailable = mixFolder.listFiles();
+                File pickedFile = pickSong(filesAvailable);
+                playMusic(pickedFile, 30000L);
+                return true;
+            } catch (IOException e) {
+                logger.error("Can't play cocorico.", e);
+                return false;
+            }
+        } else {
             return false;
+        }
+    }
+
+    @PreDestroy
+    private void tearDown() {
+        if (musicEnabled) {
+            stop();
         }
     }
 
     /**
      * This methods returns true if music is playing
+     *
      * @return true if music is playing
      */
     public boolean isPlaying() {
+        logger.info("status of player is request, current process is null : {} and is alive : {}",
+                currentMusicProcess != null,
+                currentMusicProcess != null && currentMusicProcess.isAlive());
         return (currentMusicProcess != null && currentMusicProcess.isAlive());
     }
